cleanup: move internal `npmAddScriptSync` function
and finish removing `lib/utils`

with zero test updates since this change should *not* affect
any actual behavior
