+++ lib/lib.js
@@ -10,12 +10,11 @@ const execaDefault = require('execa');
 // default fs object
 const fsExtra = require('fs-extra');
 
+const jsonfile = require('jsonfile');
+
 // Internal imports:
 
 const normalizedOptions = require('./normalized-options');
-
-// FUTURE TBD remaining module in utils to be factored out:
-const { npmAddScriptSync } = require('./utils');
 
 // Imports from templates
 
@@ -44,6 +43,22 @@ const renderTemplateIfValid = (fs, root, template, templateArgs) => {
   return fs.ensureDir(baseDir).then(() =>
     fs.outputFile(filename, template.content(templateArgs))
   );
+};
+
+// FUTURE TBD make this asynchronous and possibly more functional:
+const npmAddScriptSync = (packageJsonPath, script, fs) => {
+  try {
+    var packageJson = jsonfile.readFileSync(packageJsonPath, { fs });
+    if (!packageJson.scripts) packageJson.scripts = {};
+    packageJson.scripts[script.key] = script.value;
+    jsonfile.writeFileSync(packageJsonPath, packageJson, { fs, spaces: 2 });
+  } catch (e) {
+    if (/ENOENT.*package.json/.test(e.message)) {
+      throw new Error(`The package.json at path: ${packageJsonPath} does not exist.`);
+    } else {
+      throw e;
+    }
+  }
 };
 
 const generateWithNormalizedOptions = ({
