--- lib/utils/index.js
@@ -1,5 +1,0 @@
-const npmAddScriptSync = require('./npmAddScriptSync');
-
-module.exports = {
-  npmAddScriptSync,
-};
